# mobile/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Anupalan API",
    "summary": "Compliance engine for packaged commodities in India",
    "description": "Advisory pre-audit tool. Output is not a certification and carries no legal force. The rule pack is an engineering transcription pending legal review.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/products/listings/check": {
      "post": {
        "tags": ["products"],
        "summary": "Check a CSV of marketplace listings (Mode B)",
        "description": "Run presence and format rules over every listing in the CSV (FR-10).\n\nGated on ``PRODUCT_READ`` rather than ``PRODUCT_WRITE``: the check persists nothing, creates\nnothing, and reads only text the caller supplied in the request body. Requiring the write\npermission would lock the flagship Mode B feature away from the analyst role that exists to\nrun it, for a call that leaves no trace.\n\n``as_of`` is today. Unlike a scan \u2014 which is judged at its capture date, because FR-04 lets a\nphotograph sit in an offline queue for days \u2014 a pasted listing has no capture date. It is a\nlive listing being checked now, so now is the honest date to judge it at.",
        "operationId": "check_listings_v1_products_listings_check_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkListingIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkListingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/products": {
      "get": {
        "tags": ["products"],
        "summary": "List and search the product catalogue",
        "description": "This org's products, alphabetically (FR-03, FR-09).\n\nAlphabetical rather than newest-first because the caller is a person looking for a product they\nalready have in mind, in a picker. Recency is the right order for a history of events and the\nwrong one for a catalogue.",
        "operationId": "list_products_v1_products_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free text over name and brand",
              "title": "Q"
            },
            "description": "Free text over name and brand"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 50
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductPageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BBoxOut": {
        "properties": {
          "x": {
            "type": "number",
            "title": "X"
          },
          "y": {
            "type": "number",
            "title": "Y"
          },
          "width": {
            "type": "number",
            "title": "Width"
          },
          "height": {
            "type": "number",
            "title": "Height"
          }
        },
        "type": "object",
        "required": ["x", "y", "width", "height"],
        "title": "BBoxOut",
        "description": "Evidence rectangle on the rectified image, in pixels (FR-05: tappable)."
      },
...
      "BulkListingIn": {
        "properties": {
          "csv": {
            "type": "string",
            "minLength": 1,
            "title": "Csv",
            "description": "CSV text. Needs a header with a listing text column \u2014 one of `listing_text`, `text`, `description`, `listing` \u2014 plus optional `listing_id`/`sku`, `url`, and any profile columns (`is_imported`, `category_code`, `net_qty_value`, ...)."
          },
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProfileIn"
              },
              {
                "type": "null"
              }
            ],
            "description": "Defaults for rows that do not carry their own profile columns."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["csv"],
        "title": "BulkListingIn",
        "description": "A pasted CSV of marketplace listings (FR-10).\n\nA single text field rather than a multipart upload: \"paste or upload\" are the same payload\nonce the client has read the file, and one content type is one thing to get right in the\ngenerated mobile client."
      },
      "BulkListingOut": {
        "properties": {
          "rulepack_version": {
            "type": "string",
            "title": "Rulepack Version",
            "description": "Every verdict here was issued under this pack (CLAUDE.md \u00a73.6)"
          },
          "as_of": {
            "type": "string",
            "title": "As Of"
          },
          "scale": {
            "type": "string",
            "const": "none",
            "title": "Scale",
            "description": "A listing carries no physical scale, so every metric and geometry rule in every row is NOT_ASSESSABLE. Stated on the response so a client cannot render these verdicts as though they came from a measured photograph.",
            "default": "none"
          },
          "summary": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Summary"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ListingRowOut"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": ["rulepack_version", "as_of", "summary"],
        "title": "BulkListingOut",
        "description": "The whole upload."
      },
...
      "FindingOut": {
        "properties": {
          "finding_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finding Id",
            "description": "The stored row. Stable within an evaluation revision and replaced wholesale by the next one, because a recompute writes new rows rather than mutating these. **Null when the finding was computed rather than stored** \u2014 the bulk listing check (FR-10) judges text that was never a scan, so there is no evidence row to point at, and that is the same reason such a finding can never be corrected or reported on."
          },
          "rule_id": {
            "type": "string",
            "title": "Rule Id"
          },
          "verdict": {
            "type": "string",
            "enum": ["PASS", "FAIL", "BORDERLINE", "NOT_ASSESSABLE"],
            "title": "Verdict"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "citation": {
            "type": "string",
            "title": "Citation",
            "description": "The sub-rule, verbatim from the pack"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "observed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed"
          },
          "required": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required"
          },
          "band": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Band",
            "description": "The uncertainty band printed on a BORDERLINE verdict, e.g. 1.80-2.30. Printing it is what stops a borderline reading being read as an accusation."
          },
          "field_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Field Codes"
          },
          "bbox": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BBoxOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          }
        },
        "type": "object",
        "required": ["rule_id", "verdict", "severity", "citation"],
        "title": "FindingOut",
        "description": "One rule's verdict."
      },
...
      "FindingsSummary": {
        "properties": {
          "pass": {
            "type": "integer",
            "title": "Pass",
            "description": "PASS count",
            "default": 0
          },
          "fail": {
            "type": "integer",
            "title": "Fail",
            "default": 0
          },
          "borderline": {
            "type": "integer",
            "title": "Borderline",
            "default": 0
          },
          "na": {
            "type": "integer",
            "title": "Na",
            "description": "NOT_ASSESSABLE \u2014 could not be checked from this image",
            "default": 0
          },
          "not_applicable": {
            "type": "integer",
            "title": "Not Applicable",
            "description": "Rules that did not apply to this product \u2014 not a verdict",
            "default": 0
          }
        },
        "type": "object",
        "title": "FindingsSummary",
        "description": "Counts per bucket."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ListingRowOut": {
        "properties": {
          "row_number": {
            "type": "integer",
            "title": "Row Number",
            "description": "1-based line in the uploaded file, header included \u2014 what a spreadsheet shows"
          },
          "listing_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listing Id"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url",
            "description": "Recorded as provenance and never fetched. An endpoint that requested caller-supplied URLs would be a server-side request forgery with a CSV interface."
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Why this row could not be checked. Null when it was."
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FindingsSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingOut"
            },
            "type": "array",
            "title": "Findings"
          },
          "not_applicable_rule_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Not Applicable Rule Ids"
          }
        },
        "type": "object",
        "required": ["row_number"],
        "title": "ListingRowOut",
        "description": "One row's verdict."
      },
...
      "ProductOut": {
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "org_id": {
            "type": "string",
            "format": "uuid",
            "title": "Org Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "category_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Code"
          },
          "gtin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gtin"
          },
          "is_imported": {
            "type": "boolean",
            "title": "Is Imported",
            "default": false
          },
          "pack_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pack Type"
          },
          "surface": {
            "type": "string",
            "title": "Surface",
            "default": "printed"
          },
          "net_qty_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Qty Value"
          },
          "net_qty_unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Qty Unit"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": ["product_id", "org_id", "name", "created_at"],
        "title": "ProductOut",
        "description": "One catalogue entry.\n\nFlat rather than a nested profile, and the flatness is the honest shape. A scan's ``profile`` is\na complete set of answers because FR-03 makes the operator supply the missing ones; a catalogue\nrow is whatever has been recorded about a product so far. Nesting it would imply the two are\ninterchangeable and would need this endpoint to invent the fields a catalogue does not hold."
      },
      "ProductPageOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProductOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Opaque \u2014 see routers/pagination.py. Null on the last page."
          }
        },
        "type": "object",
        "title": "ProductPageOut",
        "description": "A page of catalogue entries, alphabetical."
      },
      "ProfileIn": {
        "properties": {
          "is_imported": {
            "type": "boolean",
            "title": "Is Imported",
            "default": false
          },
          "surface": {
            "type": "string",
            "title": "Surface",
            "default": "printed"
          },
          "qty_basis": {
            "type": "string",
            "enum": ["weight_or_volume", "length_area_or_number"],
            "title": "Qty Basis",
            "default": "weight_or_volume"
          },
          "channel": {
            "type": "string",
            "enum": ["retail", "ecommerce"],
            "title": "Channel",
            "default": "retail"
          },
          "net_qty_in_g_or_ml": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Qty In G Or Ml",
            "description": "Net quantity normalised to g or ml \u2014 the Table-I key"
          },
          "pdp_area_cm2": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdp Area Cm2",
            "description": "Principal display panel area \u2014 the Table-II key"
          },
          "net_qty_value": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Qty Value"
          },
          "net_qty_unit": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Qty Unit"
          },
          "pack_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Pack Type"
          },
          "category_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Code"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ProfileIn",
        "description": "The product context that decides which rules apply (FR-03).\n\nMirrors ``services.rules.types.Profile``. Field names are part of the rule pack's contract \u2014\npacks address them as ``profile.is_imported``, ``profile.net_qty_in_g_or_ml`` \u2014 so renaming\none here breaks a pack, not just a client."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}